+++ src/interface.ts
@@ -11,10 +11,33 @@ export interface IResponse {
     requested_params?: IParams;
 }
 
+export interface IEnterpriseResponse {
+    status: 'error' | 'success' | string;
+    result: IEnterpriseResult[];
+    execution_time: string;
+}
+
+export interface IEnterpriseResult {
+    songs: IEnterpriseGuess[];
+    offset: string;
+}
+
 export interface IGuess {
     status: 'error' | 'success' | string;
     error?: IError;
     result: IGuessResult | null;
+}
+
+export interface IEnterpriseGuess {
+    score: number;
+    artist: string;
+    title: string;
+    album: string;
+    release_date: string;
+    label: string;
+    isrc: string;
+    upc: string;
+    song_link: string;
 }
 
 interface IGuessResult {
