+++ tsconfig.json
@@ -6,11 +6,6 @@
         "outDir": "./lib",
         "strict": true
     },
-    "include": [
-        "src"
-    ],
-    "exclude": [
-        "node_modules",
-        "examples"
-    ]
+    "include": ["src"],
+    "exclude": ["node_modules", "examples"]
 }
